requirements.txt
accelerate==0.28.0
aio-pika==9.4.1
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiormq==6.8.0
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
Brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
contexttimer==0.3.3
contourpy==1.2.1
cryptography==42.0.5
cuda-python==12.4.0
cycler==0.12.1
cymem==2.0.8
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
dill==0.3.8
distlib==0.3.8
einops==0.7.0
environs==9.5.0
executing==2.0.1
fairscale==0.4.4
filelock==3.13.3
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
gevent==24.2.1
geventhttpclient==2.0.2
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.60.0
huggingface-hub==0.22.2
humanfriendly==10.0
hydra-core==1.3.2
identify==2.5.35
idna==3.6
imageio==2.34.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaggle==1.6.11
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
lightning==2.2.2
lightning-utilities==0.11.2
loguru==0.7.2
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
minio==7.2.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.17.1
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
optimum==1.18.0
packaging==24.0
pamqp==3.3.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.20.0
portalocker==2.8.2
pre-commit==3.7.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==3.20.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-settings==2.2.1
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pymilvus==2.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-rapidjson==1.16
python-slugify==8.0.4
pytorch-lightning==2.2.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.4.12
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
triton==2.2.0
tritonclient==2.44.0
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
widgetsnbextension==4.0.10
xxhash==3.4.1
yarl==1.9.4
zope.event==5.0
zope.interface==6.2
